# docs/api/openapi.yaml
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    PasswordChangeIn:
      description: '改口令输入。``old_password`` 只校验非空 —— 强度是**当前**口令的事，

        强制它满足现行强度规则会让「弱口令时代设的旧口令」永远改不了。'
      properties:
        new_password:
          maxLength: 1024
          minLength: 1
          title: New Password
          type: string
        old_password:
          maxLength: 1024
          minLength: 1
          title: Old Password
          type: string
      required:
      - old_password
      - new_password
      title: PasswordChangeIn
      type: object
    PasswordIn:
      description: '新口令输入。


        强度校验放在**这里**（pydantic 字段校验器）而不是 ``hash_password`` 里：

        在路由函数体里抛 ``ValueError`` 会变成 500，那是服务器错误不是客户端错误。

        走校验器则自动 422，且 ``ctx`` 里的异常对象由 ``jsonable_errors`` 兜住

        （见 ``app/core/errors.py`` 里那个 422 序列化的坑）。'
      properties:
        password:
          maxLength: 1024
          minLength: 1
          title: Password
          type: string
      required:
      - password
      title: PasswordIn
      type: object
...
    SessionOut:
      description: '登录/创建口令成功后下发的会话令牌。


        令牌是 HMAC 签名（不是随机串），因此服务端无需会话表；它的有效期由

        签发时的 TTL 决定，改口令会让它立刻失效。**只在响应里出现，不写日志。**'
      properties:
        expires_at:
          format: date-time
          title: Expires At
          type: string
        session_token:
          title: Session Token
          type: string
      required:
      - session_token
      - expires_at
      title: SessionOut
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  description: '学习工作台 Strayt 服务端。


    单用户免账号：所有 `/api/v1` 端点需 `Authorization: Bearer <访问令牌>`。

    唯一豁免是 `/health`（存活探针）。


    错误体统一为 `{ "error": { "code", "message", "detail?" } }`，`code` 是稳定契约，客户端据此分支。'
  title: Strayt 学习工作台
  version: 0.1.0
openapi: 3.1.0
paths:
  /api/v1/auth/login:
    post:
      description: 口令换会话令牌。
      operationId: login_api_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Login
      tags:
      - 访问口令登录
  /api/v1/auth/password:
    post:
      description: '改口令。鉴权由 ``router2`` 在 ``main.py`` 挂载时用 ``AuthDep`` 施加。


        改完立刻返回**新**会话令牌：签名密钥由口令哈希派生，换口令等于换密钥，

        客户端手上那个旧令牌当场作废。不回传的话用户会被自己踢下线，得重新输一遍。'
      operationId: change_password_api_v1_auth_password_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordChangeIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Change Password
      tags:
      - 访问口令登录
  /api/v1/auth/setup:
    post:
      description: '首次创建口令。已设置过则 409 —— 不给「静默覆盖」留口子。


        口令强度已由 ``PasswordIn`` 的字段校验器把关（弱口令 → 422），这里

        ``hash_password`` 只会因内部不一致抛错，那确实是 500。'
      operationId: setup_password_api_v1_auth_setup_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Setup Password
      tags:
      - 访问口令登录
...